# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BusinessShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /scans/{id}/report:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a compliance report for a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ─── Policies ────────────────────────────────────────────────────────────────
...
  /policies/generate:
    post:
      operationId: generatePolicy
      tags: [policies]
      summary: Generate a policy document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyGenerateInput"
      responses:
        "201":
          description: Generated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a policy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: Delete a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Report:
      type: object
      required: [id, userId, scanId, title, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        scanId:
          type: integer
        websiteId:
          type: ["integer", "null"]
        title:
          type: string
        status:
          type: string
          enum: [generating, ready, failed]
        executiveSummary:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        complianceScore:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Policies ──────────────────────────────────────────────────────────────
    Policy:
      type: object
      required: [id, userId, type, title, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        websiteId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [privacy_policy, cookie_policy, accessibility_statement, terms_of_service]
        title:
          type: string
        content:
          type: string
        websiteUrl:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PolicyGenerateInput:
      type: object
      required: [type, companyName, websiteUrl]
      properties:
        type:
          type: string
          enum: [privacy_policy, cookie_policy, accessibility_statement, terms_of_service]
        companyName:
          type: string
        websiteUrl:
          type: string
        websiteId:
          type: ["integer", "null"]
        additionalContext:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...